+++ stroom-core-client-widget/src/main/java/edu/ycp/cs/dh/acegwt/public/ace/ace.js
@@ -15702,13 +15702,7 @@ var Gutter = function(parentEl) {
             if (rowInfo.text.indexOf(annoText) === -1)
                 rowInfo.text.push(annoText);
 
-            var type = annotation.type;
-            if (type == "error")
-                rowInfo.className = " ace_error";
-            else if (type == "warning" && rowInfo.className != " ace_error")
-                rowInfo.className = " ace_warning";
-            else if (type == "info" && (!rowInfo.className))
-                rowInfo.className = " ace_info";
+            rowInfo.className = " " + annotation.type;
         }
     };