+++ stroom-core-server/src/main/java/stroom/datasource/DataSourceProviderRegistryImpl.java
@@ -8,7 +8,6 @@ import stroom.node.server.StroomPropertyService;
 import stroom.query.api.v2.DocRef;
 import stroom.security.SecurityContext;
 import stroom.servicediscovery.ServiceDiscoverer;
-import stroom.servlet.HttpServletRequestHolder;
 import stroom.util.spring.StroomBeanStore;
 import stroom.util.spring.StroomScope;
 
@@ -22,21 +21,15 @@ public class DataSourceProviderRegistryImpl implements DataSourceProviderRegistr
 
     private static final Logger LOGGER = LoggerFactory.getLogger(DataSourceProviderRegistryImpl.class);
 
-    public static final String PROP_KEY_SERVICE_DISCOVERY_ENABLED = "stroom.serviceDiscovery.enabled";
+    private static final String PROP_KEY_SERVICE_DISCOVERY_ENABLED = "stroom.serviceDiscovery.enabled";
 
-    private final SecurityContext securityContext;
-    private final StroomPropertyService stroomPropertyService;
     private final DataSourceProviderRegistry delegateDataSourceProviderRegistry;
 
     @SuppressWarnings("unused")
     @Inject
     public DataSourceProviderRegistryImpl(final SecurityContext securityContext,
                                           final StroomPropertyService stroomPropertyService,
-                                          final StroomBeanStore stroomBeanStore,
-                                          final HttpServletRequestHolder httpServletRequestHolder) {
-        this.securityContext = securityContext;
-        this.stroomPropertyService = stroomPropertyService;
-
+                                          final StroomBeanStore stroomBeanStore) {
         boolean isServiceDiscoveryEnabled = stroomPropertyService.getBooleanProperty(
                 PROP_KEY_SERVICE_DISCOVERY_ENABLED,
                 false);
@@ -46,14 +39,12 @@ public class DataSourceProviderRegistryImpl implements DataSourceProviderRegistr
             LOGGER.debug("Using service discovery for service lookup");
             delegateDataSourceProviderRegistry = new ServiceDiscoveryDataSourceProviderRegistry(
                     securityContext,
-                    serviceDiscoverer,
-                    httpServletRequestHolder);
+                    serviceDiscoverer);
         } else {
             LOGGER.debug("Using local services");
             delegateDataSourceProviderRegistry = new SimpleDataSourceProviderRegistry(
                     securityContext,
-                    stroomPropertyService,
-                    httpServletRequestHolder);
+                    stroomPropertyService);
         }
     }