+++ stroom-dashboard-server/src/main/java/stroom/dashboard/server/ActiveQueries.java
@@ -22,24 +22,30 @@ import stroom.dashboard.shared.DashboardQueryKey;
 import stroom.datasource.DataSourceProviderRegistry;
 import stroom.query.api.v2.DocRef;
 import stroom.query.api.v2.QueryKey;
+import stroom.security.SecurityContext;
+import stroom.security.SecurityHelper;
+import stroom.security.UserTokenUtil;
 
 import java.util.Iterator;
 import java.util.Map.Entry;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 
-public class ActiveQueries {
+class ActiveQueries {
     private static final Logger LOGGER = LoggerFactory.getLogger(ActiveQueries.class);
 
     private final ConcurrentHashMap<DashboardQueryKey, ActiveQuery> activeQueries = new ConcurrentHashMap<>();
 
+    private final SecurityContext securityContext;
     private final DataSourceProviderRegistry dataSourceProviderRegistry;
 
-    public ActiveQueries(final DataSourceProviderRegistry dataSourceProviderRegistry) {
+    ActiveQueries(final DataSourceProviderRegistry dataSourceProviderRegistry,
+                  final SecurityContext securityContext) {
         this.dataSourceProviderRegistry = dataSourceProviderRegistry;
+        this.securityContext = securityContext;
     }
 
-    public void destroyUnusedQueries(final Set<DashboardQueryKey> keys) {
+    void destroyUnusedQueries(final Set<DashboardQueryKey> keys) {
         // Kill off any searches that are no longer required by the UI.
         Iterator<Entry<DashboardQueryKey, ActiveQuery>> iterator = activeQueries.entrySet().iterator();
         while (iterator.hasNext()) {
@@ -47,30 +53,36 @@ public class ActiveQueries {
             final DashboardQueryKey queryKey = entry.getKey();
             final ActiveQuery activeQuery = entry.getValue();
             if (keys == null || !keys.contains(queryKey)) {
-                // Terminate the associated search task.
-                Boolean success = dataSourceProviderRegistry.getDataSourceProvider(activeQuery.getDocRef())
-                        .map(provider -> provider.destroy(new QueryKey(queryKey.getUuid())))
-                        .orElseGet(() -> {
-                            LOGGER.warn("Unable to destroy query with key {} as provider {} cannot be found",
-                                    queryKey.getUuid(),
-                                    activeQuery.getDocRef().getType());
-                            return Boolean.TRUE;
-                        });
+                try (final SecurityHelper securityHelper = SecurityHelper.asUser(securityContext, UserTokenUtil.create(activeQuery.getUserId(), null))) {
+                    // Terminate the associated search task.
+                    Boolean success = dataSourceProviderRegistry.getDataSourceProvider(activeQuery.getDocRef())
+                            .map(provider -> provider.destroy(new QueryKey(queryKey.getUuid())))
+                            .orElseGet(() -> {
+                                LOGGER.warn("Unable to destroy query with key {} as provider {} cannot be found",
+                                        queryKey.getUuid(),
+                                        activeQuery.getDocRef().getType());
+                                return Boolean.TRUE;
+                            });
 
-                if (Boolean.TRUE.equals(success)) {
-                    // Remove the collector from the available searches as it is no longer required by the UI.
-                    iterator.remove();
+                    if (Boolean.TRUE.equals(success)) {
+                        // Remove the collector from the available searches as it is no longer required by the UI.
+                        iterator.remove();
+                    }
                 }
             }
         }
     }
 
-    public ActiveQuery getExistingQuery(final DashboardQueryKey queryKey) {
+    ActiveQuery getExistingQuery(final DashboardQueryKey queryKey) {
         return activeQueries.get(queryKey);
     }
 
-    public ActiveQuery addNewQuery(final DashboardQueryKey queryKey, final DocRef docRef) {
-        final ActiveQuery activeQuery = new ActiveQuery(docRef);
+    ActiveQuery addNewQuery(final DashboardQueryKey queryKey, final DocRef docRef) {
+        final String userId = securityContext.getUserId();
+        if (userId == null) {
+            throw new RuntimeException("No user is currently logged in");
+        }
+        final ActiveQuery activeQuery = new ActiveQuery(docRef, userId);
         final ActiveQuery existing = activeQueries.put(queryKey, activeQuery);
         if (existing != null) {
             throw new RuntimeException(